requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
hume-tada
transformers<5
